+++ src/sokoenginepy/io/collection.py
@@ -1,16 +1,25 @@
 from __future__ import annotations
 
+import io
 from pathlib import Path
-from typing import TYPE_CHECKING, List, Optional, Union
+from typing import List, Optional, Union
 
+from ..game import Tessellation
 from .puzzle import Puzzle
-
-if TYPE_CHECKING:
-    from ..game import Tessellation
 
 
 class Collection:
-    """Collection of one or more game puzzles."""
+    """
+    Collection of one or more game puzzles.
+
+    Attributes:
+        title (str): Collection title
+        author (str): Collection author
+        created_at (str): datetime string of unspecified format (it is not parsed)
+        updated_at (str): datetime string of unspecified format (it is not parsed)
+        notes (List[str]): collection notes
+        puzzles (List[Puzzle]): collection puzzles
+    """
 
     def __init__(
         self,
@@ -28,47 +37,87 @@ class Collection:
         self.puzzles: List[Puzzle] = []
 
     def load(
-        self, path: Union[str, Path], tessellation_hint: Optional[Tessellation] = None
+        self,
+        src: Union[
+            str,
+            Path,
+            io.BufferedReader,
+            io.TextIOWrapper,
+            io.FileIO,
+            io.StringIO,
+            io.BytesIO,
+        ],
+        tessellation_hint: Tessellation = Tessellation.SOKOBAN,
     ):
         """
-        Loads collection from ``path``.
+        Loads collection from ``src``.
 
-        Loader supports SokobanYASC .sok format, but will happily try to load older
-        similar, textual sokoban files (usually with extensions ``.txt``, ``.tsb`` and
+        Loader supports SokobanYASC .sok format, but will happily try to load older,
+        similar textual sokoban files (usually with extensions ``.txt``, ``.tsb`` or
         ``.hsb``).
 
         Arguments:
-            path: source file path
+            src: source file path or input stream object
             tessellation_hint: If puzzles in file don't specify their game tessellation
                 assume this value.
         """
         from .sok_file_format import SOKFileFormat
 
-        with open(path, "r") as f:
-            SOKFileFormat.read(
-                f, self, tessellation_hint or self._extension_to_tessellation_hint(path)
-            )
+        if isinstance(src, (str, Path)):
+            with open(src, "r") as f:
+                SOKFileFormat.read(f, self, tessellation_hint)
+        else:
+            SOKFileFormat.read(src, self, tessellation_hint)
 
-    def save(self, path: Union[str, Path]):
+    def loads(
+        self,
+        data: Union[str, bytes],
+        tessellation_hint: Tessellation = Tessellation.SOKOBAN,
+    ):
         """
-        Saves collection to file at ``path`` in SokobanYASC .sok format.
+        Loads collections from ``data``.
+
+        Arguments:
+            data: raw collection data
+            tessellation_hint: If puzzles in file don't specify their game tessellation
+                assume this value.
+        """
+        if isinstance(data, bytes):
+            data = data.decode(encoding="utf-8")
+        f = io.StringIO(data)
+        self.load(f, tessellation_hint)
+
+    def dump(
+        self,
+        dst: Union[
+            str,
+            Path,
+            io.BufferedWriter,
+            io.TextIOWrapper,
+            io.FileIO,
+            io.StringIO,
+            io.BytesIO,
+        ],
+    ):
+        """
+        Saves collection to ``dst`` in SokobanYASC .sok format.
 
         Note:
-            File can have any kind of extension, not necessary ``.sok``.
+            Doesn't care about file extension if ``dst`` is path to file.
+
+        Arguments:
+            dst: Path to destination file or destination stream object.
         """
         from .sok_file_format import SOKFileFormat
 
-        with open(path, "w") as f:
-            SOKFileFormat.write(self, f)
-
-    @staticmethod
-    def _extension_to_tessellation_hint(path: Union[str, Path]) -> Tessellation:
-        from ..game import Tessellation
-
-        file_extension = Path(path).suffix
-        if file_extension == ".tsb":
-            return Tessellation.TRIOBAN
-        elif file_extension == ".hsb":
-            return Tessellation.HEXOBAN
+        if isinstance(dst, (str, Path)):
+            with open(dst, "w") as f:
+                SOKFileFormat.write(self, f)
         else:
-            return Tessellation.SOKOBAN
+            SOKFileFormat.write(self, dst)
+
+    def dumps(self) -> str:
+        """Saves collection to `str`."""
+        out = io.StringIO()
+        self.dump(out)
+        return out.getvalue()