+++ src/sokoenginepyext/export_io_collection.cpp
@@ -1,11 +1,20 @@
 #include "sokoenginepyext.hpp"
 
+#include <sstream>
+
 using sokoengine::game::Tessellation;
 using sokoengine::io::Collection;
 using sokoengine::io::Puzzles;
 using sokoengine::io::Strings;
+using std::istream;
+using std::istringstream;
+using std::ostream;
+using std::ostringstream;
 using std::string;
 using std::filesystem::path;
+
+void stream_load(Collection &self, py::object &stream, Tessellation tessellation_hint);
+void stream_dump(const Collection &self, py::object &stream);
 
 void export_io_collection(py::module &m) {
   auto pyCollection = py::class_<Collection>(m, "Collection");
@@ -79,12 +88,96 @@ void export_io_collection(py::module &m) {
     }
   );
 
-  pyCollection.def("load", py::overload_cast<const path &>(&Collection::load))
-    .def("load", py::overload_cast<const path &, Tessellation>(&Collection::load))
-    .def("load", py::overload_cast<const string &>(&Collection::load))
-    .def("load", py::overload_cast<const string &, Tessellation>(&Collection::load));
+  pyCollection
+    .def(
+      "load",
+      py::overload_cast<const path &, Tessellation>(&Collection::load),
+      py::arg("src").none(false),
+      py::arg("tessellation_hint") = Tessellation::SOKOBAN
+    )
+    .def(
+      "load",
+      py::overload_cast<const string &, Tessellation>(&Collection::load),
+      py::arg("src").none(false),
+      py::arg("tessellation_hint") = Tessellation::SOKOBAN
+    )
+    .def(
+      "load",
+      py::overload_cast<Collection &, py::object &, Tessellation>(&stream_load),
+      py::arg("src").none(false),
+      py::arg("tessellation_hint") = Tessellation::SOKOBAN
+    );
+
+  pyCollection.def(
+    "loads",
+    py::overload_cast<const string &, Tessellation>(&Collection::loads),
+    py::arg("data").none(false),
+    py::arg("tessellation_hint") = Tessellation::SOKOBAN
+  );
 
   pyCollection
-    .def("save", py::overload_cast<const path &>(&Collection::save, py::const_))
-    .def("save", py::overload_cast<const string &>(&Collection::save, py::const_));
+    .def(
+      "dump",
+      py::overload_cast<const path &>(&Collection::dump, py::const_),
+      py::arg("dst").none(false)
+    )
+    .def(
+      "dump",
+      py::overload_cast<const string &>(&Collection::dump, py::const_),
+      py::arg("dst").none(false)
+    )
+    .def(
+      "dump",
+      py::overload_cast<const Collection &, py::object &>(&stream_dump),
+      py::arg("dst").none(false)
+    );
+
+  pyCollection.def("dumps", &Collection::dumps);
+}
+
+void stream_load(Collection &self, py::object &stream, Tessellation tessellation_hint) {
+  // Assume stream is Python stream object (io.TextIOWrapper, io.StringIO, ...)
+  // and read all lines from it into memory.
+  //
+  // This is quite inefficient, but only way to preserve Python Collection.load API
+  //
+  // Since collections are expected to be only few MiB in size, we don't care about
+  // inefficiency here.
+  string data;
+
+  {
+    stream.attr("seek")(0, 0);
+    ostringstream conv;
+    for (auto &line : stream) {
+      if (py::isinstance<py::bytes>(line)) {
+        conv << line.attr("decode")("utf-8");
+      } else {
+        conv << line;
+      }
+    }
+    data = conv.str();
+  }
+
+  istringstream input(data);
+  self.load(input);
+}
+
+void stream_dump(const Collection &self, py::object &stream) {
+  string data;
+
+  ostringstream conv;
+  self.dump(conv);
+
+  py::str stream_type_name = py::str(stream.get_type().attr("__name__"));
+
+  if (
+    stream_type_name.equal(py::str("BufferedWriter"))
+    || stream_type_name.equal(py::str("BytesIO"))
+    || stream_type_name.equal(py::str("FileIO"))
+  ) {
+    std::cout << py::str(conv.str()).attr("encode")("utf-8") << std::endl;
+    stream.attr("write")(py::str(conv.str()).attr("encode")("utf-8"));
+  } else {
+    stream.attr("write")(conv.str());
+  }
 }